requirements.txt
aiogram==3.15.0
SQLAlchemy[asyncio]==2.0.36
aiosqlite==0.20.0
yookassa==3.4.0
python-dotenv==1.0.1
aiohttp==3.11.11
